refactor move Levels to util
